fix unit tests

+++ resource-validator/src/main/scala/com/broadinstitute/dsp/resourceValidator/DeletedDiskChecker.scala
@@ -26,18 +26,19 @@ object DeletedDiskChecker {
       ): F[Option[Disk]] =
         for {
           diskOpt <- deps.googleDiskService.getDisk(disk.googleProject, disk.zone, disk.diskName)
-          _ <- if (!isDryRun) {
-            if (disk.formattedBy.getOrElse(None) == "GALAXY") {
-              val dataDiskName = disk.diskName
-              val postgresDiskName = DiskName(s"${dataDiskName.value}-gxy-postres-disk")
-              diskOpt.traverse { _ =>
-                List(postgresDiskName, dataDiskName).parTraverse(dn =>
-                  deps.googleDiskService.deleteDisk(disk.googleProject, disk.zone, dn)
-                )
-              }
-            } else
-              diskOpt.traverse(_ => deps.googleDiskService.deleteDisk(disk.googleProject, disk.zone, disk.diskName))
-          } else F.pure(None)
+          _ <-
+            if (!isDryRun) {
+              if (disk.formattedBy.getOrElse(None) == "GALAXY") {
+                val dataDiskName = disk.diskName
+                val postgresDiskName = DiskName(s"${dataDiskName.value}-gxy-postres-disk")
+                diskOpt.traverse { _ =>
+                  List(postgresDiskName, dataDiskName).parTraverse(dn =>
+                    deps.googleDiskService.deleteDisk(disk.googleProject, disk.zone, dn)
+                  )
+                }
+              } else
+                diskOpt.traverse(_ => deps.googleDiskService.deleteDisk(disk.googleProject, disk.zone, disk.diskName))
+            } else F.pure(None)
         } yield diskOpt.map(_ => disk)
     }
 }
